Create BookList component
