+++ src/components/BookList.js
@@ -1,0 +1,23 @@
+import { useSelector } from 'react-redux';
+import Book from './Book';
+
+const BookList = () => {
+  const bookStore = useSelector((state) => state.booksReducer);
+
+  return (
+    <div>
+      {
+        bookStore.map((book) => (
+          <Book
+            title={book.title}
+            author={book.author}
+            key={book.id}
+            id={book.id}
+          />
+        ))
+      }
+    </div>
+  );
+};
+
+export default BookList;
